feat : 로그아웃이나 회원탈퇴시 sse 연결이 끊어지도록 추가
